+++ src/types.ts
@@ -1,33 +1,11 @@
 import NanoEvents from "nanoevents";
+import { Data, DebugConfig, Delta, Events } from "./core";
 import { Connectivity, NetworkAdapter } from "./network";
 
 export type CreateParams<Domain> = {
   network: NetworkAdapter<Domain>;
   onError?: (error: Error) => void;
   window?: Window;
-};
-
-export type Data<Domain> = {
-  [k in keyof Domain]: {
-    [id: string]: { revision: string; value: Domain[k] } | undefined;
-  }
-};
-
-export type Delta<Domain, K extends keyof Domain> = (
-  prev: Domain[K]
-) => Domain[K];
-
-export type DebugConfig = {
-  failingWrites: boolean;
-  pretendOffline: boolean;
-  optimisticUIEnabled: boolean;
-};
-
-export type Events<Domain> = {
-  change: [keyof Domain, string];
-  "debug-config-changed": undefined;
-  "pending-transaction-count-changed": undefined;
-  "connectivity-changed": undefined;
 };
 
 export type FullDB<Domain> = {
