+++ src/ui/react/configure.ts
@@ -1,8 +1,9 @@
 import { create } from "../../create";
+import { Data } from "../../core";
 import { createContext } from "./context";
 import { makeHooks } from "./hooks";
 import { makeSemiControlledInput } from "./inputs";
-import { CreateParams, Data } from "../../types";
+import { CreateParams } from "../../types";
 import { ReactType } from "./types";
 
 type Configuration<Domain> = {
