+++ src/ui/react/hooks.ts
@@ -1,4 +1,5 @@
-import { DebugConfig, FullDB } from "../../types";
+import { DebugConfig } from "../../core";
+import { FullDB } from "../../types";
 import { ReactType } from "./types";
 
 type UpdaterInternal<T> = (options: {}, delta: (prev: T) => T) => void;
